chore: switch locations between data structure <-> state
