+++ promkit-widgets/src/cursor.rs
@@ -1,4 +1,4 @@
-mod len;
+pub mod len;
 use len::Len;
 
 /// A generic cursor structure for navigating and manipulating collections.
