+++ promkit-widgets/src/text/text.rs
@@ -1,0 +1,50 @@
+use promkit_core::grapheme::StyledGraphemes;
+
+use crate::cursor::Cursor;
+
+#[derive(Clone)]
+pub struct Text(Cursor<Vec<StyledGraphemes>>);
+
+impl Default for Text {
+    fn default() -> Self {
+        Self(Cursor::new(vec![], 0, false))
+    }
+}
+
+impl<T: AsRef<str>> From<T> for Text {
+    fn from(text: T) -> Self {
+        let lines: Vec<StyledGraphemes> = text
+            .as_ref()
+            .split('\n')
+            // Replace empty lines with null character to
+            // prevent them from being ignored at `style::Print`
+            .map(|line| if line.is_empty() { "\0" } else { line })
+            .map(StyledGraphemes::from)
+            .collect();
+        Self(Cursor::new(lines, 0, false))
+    }
+}
+
+impl Text {
+    /// Returns a reference to the vector of items in the listbox.
+    pub fn items(&self) -> &Vec<StyledGraphemes> {
+        self.0.contents()
+    }
+
+    /// Returns the current position of the cursor within the listbox.
+    pub fn position(&self) -> usize {
+        self.0.position()
+    }
+
+    /// Moves the cursor backward in the listbox, if possible.
+    /// Returns `true` if the cursor was successfully moved backward, `false` otherwise.
+    pub fn backward(&mut self) -> bool {
+        self.0.backward()
+    }
+
+    /// Moves the cursor forward in the listbox, if possible.
+    /// Returns `true` if the cursor was successfully moved forward, `false` otherwise.
+    pub fn forward(&mut self) -> bool {
+        self.0.forward()
+    }
+}
