+++ promkit-widgets/src/tree.rs
@@ -1,85 +1,101 @@
-mod node;
+use promkit_core::{
+    PaneFactory, crossterm::style::ContentStyle, grapheme::StyledGraphemes, pane::Pane,
+};
 
-use crate::cursor::Cursor;
+pub mod node;
+use node::Kind;
+mod tree;
+pub use tree::Tree;
 
-pub use node::{Kind, Node, Path};
-mod state;
-pub use state::State;
-
-/// A `Tree` structure that manages a collection of nodes in a hierarchical manner.
-/// It utilizes a cursor to navigate and manipulate the nodes within the tree.
+/// Represents the state of a tree structure within the application.
+///
+/// This state includes not only the tree itself but also various properties
+/// that affect how the tree is displayed and interacted with. These properties
+/// include symbols for folded and unfolded items, styles for active and inactive
+/// items, the number of lines available for rendering, and the indentation level
+/// for child items in the tree.
 #[derive(Clone)]
-pub struct Tree {
-    root: Node,
-    cursor: Cursor<Vec<Kind>>,
+pub struct State {
+    pub tree: Tree,
+
+    /// Symbol representing folded items.
+    pub folded_symbol: String,
+    /// Symbol representing unfolded items.
+    pub unfolded_symbol: String,
+
+    /// Style for the selected line.
+    pub active_item_style: ContentStyle,
+    /// Style for un-selected lines.
+    pub inactive_item_style: ContentStyle,
+
+    /// Number of lines available for rendering.
+    pub lines: Option<usize>,
+
+    /// The number of spaces used for indenting child items in the tree.
+    /// This value determines how much horizontal space is used to visually
+    /// represent the hierarchical structure of the tree. Each level of
+    /// indentation typically represents a deeper level in the tree hierarchy.
+    pub indent: usize,
 }
 
-impl Tree {
-    /// Creates a new `Tree` with a given root node.
-    ///
-    /// # Arguments
-    ///
-    /// * `root` - The root node of the tree.
-    pub fn new(root: Node) -> Self {
-        Self {
-            root: root.clone(),
-            cursor: Cursor::new(root.flatten_visibles(), 0, false),
-        }
-    }
-
-    /// Returns a vector of all nodes in the tree, represented with their depth information.
-    pub fn kinds(&self) -> Vec<Kind> {
-        self.cursor.contents().clone()
-    }
-
-    /// Returns the current position of the cursor within the tree.
-    pub fn position(&self) -> usize {
-        self.cursor.position()
-    }
-
-    /// Retrieves the data of the current node pointed by the cursor, along with its path from the root.
-    pub fn get(&self) -> Vec<String> {
-        let kind = self.cursor.contents()[self.position()].clone();
-        match kind {
-            Kind::Folded { id, path } | Kind::Unfolded { id, path } => {
-                let mut ret = self.root.get_waypoints(&path);
-                ret.push(id.to_string());
-                ret
+impl PaneFactory for State {
+    fn create_pane(&self, width: u16, height: u16) -> Pane {
+        let symbol = |kind: &Kind| -> &str {
+            match kind {
+                Kind::Folded { .. } => &self.folded_symbol,
+                Kind::Unfolded { .. } => &self.unfolded_symbol,
             }
-        }
-    }
-
-    /// Toggles the state of the current node and updates the cursor position accordingly.
-    pub fn toggle(&mut self) {
-        let path = match self.cursor.contents()[self.position()].clone() {
-            Kind::Folded { path, .. } => path,
-            Kind::Unfolded { path, .. } => path,
         };
-        self.root.toggle(&path);
-        self.cursor = Cursor::new(self.root.flatten_visibles(), self.position(), false);
-    }
 
-    /// Moves the cursor backward in the tree, if possible.
-    ///
-    /// Returns `true` if the cursor was successfully moved backward, `false` otherwise.
-    pub fn backward(&mut self) -> bool {
-        self.cursor.backward()
-    }
+        let indent = |kind: &Kind| -> usize {
+            match kind {
+                Kind::Folded { path, .. } | Kind::Unfolded { path, .. } => path.len() * self.indent,
+            }
+        };
 
-    /// Moves the cursor forward in the tree, if possible.
-    ///
-    /// Returns `true` if the cursor was successfully moved forward, `false` otherwise.
-    pub fn forward(&mut self) -> bool {
-        self.cursor.forward()
-    }
+        let id = |kind: &Kind| -> String {
+            match kind {
+                Kind::Folded { id, .. } | Kind::Unfolded { id, .. } => id.clone(),
+            }
+        };
 
-    /// Moves the cursor to the head of the tree.
-    pub fn move_to_head(&mut self) {
-        self.cursor.move_to_head()
-    }
+        let height = match self.lines {
+            Some(lines) => lines.min(height as usize),
+            None => height as usize,
+        };
 
-    /// Moves the cursor to the tail of the tree.
-    pub fn move_to_tail(&mut self) {
-        self.cursor.move_to_tail()
+        let matrix = self
+            .tree
+            .kinds()
+            .iter()
+            .enumerate()
+            .filter(|(i, _)| *i >= self.tree.position() && *i < self.tree.position() + height)
+            .map(|(i, kind)| {
+                if i == self.tree.position() {
+                    StyledGraphemes::from_str(
+                        format!("{}{}{}", symbol(kind), " ".repeat(indent(kind)), id(kind),),
+                        self.active_item_style,
+                    )
+                } else {
+                    StyledGraphemes::from_str(
+                        format!(
+                            "{}{}{}",
+                            " ".repeat(StyledGraphemes::from(symbol(kind)).widths()),
+                            " ".repeat(indent(kind)),
+                            id(kind),
+                        ),
+                        self.inactive_item_style,
+                    )
+                }
+            })
+            .fold((vec![], 0), |(mut acc, pos), item| {
+                let rows = item.matrixify(width as usize, height, 0).0;
+                if pos < self.tree.position() + height {
+                    acc.extend(rows);
+                }
+                (acc, pos + 1)
+            });
+
+        Pane::new(matrix.0, 0)
     }
 }
